casepilot-main/backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.18.0
google-api-python-client>=2.149.0
google-auth>=2.35.0
google-adk>=1.0.0
google-genai>=1.0.0
python-multipart>=0.0.9
requests>=2.31.0
